requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
Pillow==10.4.0
exifread==3.0.0
pillow-heif==0.16.0